// frontend/src/api/model/domainURL.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * URL Shortener API
 * API for shortening URLs
 * OpenAPI spec version: 1.0
 */

export interface DomainURL {
  created_at?: string;
  id?: string;
  original_url?: string;
  short_code?: string;
  updated_at?: string;
  user_id?: string;
}